DreamAnalyze/requirements.txt
flask
line-bot-sdk
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
